2023 - 03 - 06 Update readme with instructions for setting up the environment and accessing the dataset server
Detailed Description:

This commit updates the readme file to provide clear instructions for setting up the necessary environment variables and accessing the CRI Hand Datasets server. It also includes a section on setting up the environment and loading the dataset using TensorFlow Dataset API.

The readme file now includes a brief description of the dataset and a citation that should be used if referencing it. Finally, the license under which the dataset is released is also mentioned.

Overall, these changes should make it easier for users to get started with the dataset and use it for computer vision research.